Refactor functions game in that handle keyboard actions to separate game-commands file
