+++ lib/scripts/index.js
@@ -1,4 +1,4 @@
-var Game = require('./game');
+var Game = require('./game-commands');
 
 require('../stylesheets/style.css');
